Enhance/login option for seperate form (#203)
* enable form login-options in form settings

* login option make work on form submit js

* add login options checks in each and every hooks

* code cleanup

+++ includes/class-ur-user-approval.php
@@ -23,16 +23,13 @@ class UR_User_Approval {
 	 */
 	public function __construct() {
 
-		if ( 'admin_approval' !== get_option( 'user_registration_general_setting_login_options' ) ) {
-			return;
-		}
 		// -------------------- ACTIONS & FILTERS --------------------
 		// Additional checks
 		add_action( 'after_setup_theme', array( $this, 'check_status_on_page' ) );
 
 		// Handle user Sign in
-		add_action( 'user_register', array( $this, 'send_request_notification_to_admin' ) );
-		add_action( 'user_register', array( $this, 'set_user_status' ) );
+		add_action( 'user_registration_after_register_user_action', array( $this, 'set_user_status' ), 10, 3 );
+		add_action( 'user_register', array( $this, 'send_request_notification_to_admin' ), 10, 1 );
 		add_filter( 'wp_login_errors', array( $this, 'registration_completed_message' ) );
 
 		// Handle user Sign on
@@ -55,9 +52,9 @@ class UR_User_Approval {
 		add_action( 'ur_user_user_denied', array( $this, 'disconnect_user_session' ) );
 
 		// Try to hide the not approved users from any theme or plugin request in frontend
-		add_action( 'pre_get_users', array( $this, 'hide_not_approved_users_in_frontend' ) );
+		// add_action( 'pre_get_users', array( $this, 'hide_not_approved_users_in_frontend' ) );
 
-		do_action( 'ur_user_construct' );
+		// do_action( 'ur_user_construct' );
 
 	}
 
@@ -84,9 +81,12 @@ class UR_User_Approval {
 	 */
 	public function track_first_login( $user_login, $user ) {
 
-		$user_manager = new UR_Admin_User_Manager( $user );
-		$user_manager->save_first_access_flag();
+		$form_id = ur_get_form_id_by_userid( $user->ID );
 
+		if ( 'admin_approval' === ur_get_single_post_meta( $form_id, 'user_registration_form_setting_login_options', get_option( 'user_registration_general_setting_login_options', 'default' ) ) ) {
+			$user_manager = new UR_Admin_User_Manager( $user );
+			$user_manager->save_first_access_flag();
+		}
 	}
 
 	/**
@@ -101,7 +101,9 @@ class UR_User_Approval {
 	 */
 	public function send_notification_to_user_about_status_changing( $status, $user_id, $alert_user ) {
 
-		if ( ! $alert_user ) {
+		$form_id = ur_get_form_id_by_userid( $user_id );
+
+		if ( ! $alert_user && 'admin_approval' !== ur_get_single_post_meta( $form_id, 'user_registration_form_setting_login_options', get_option( 'user_registration_general_setting_login_options', 'default' ) ) ) {
 			return;
 		}
 
@@ -129,10 +131,11 @@ class UR_User_Approval {
 	 */
 	public function send_request_notification_to_admin( $user_id ) {
 
-		// If the user is created by admin or if the admin alert is disabled, doesn't send the email to the admin
-		if ( $this->is_admin_creation_process() ) {
-			return;
-		}
+			// If the user is created by admin or if the admin alert is disabled, doesn't send the email to the admin
+			if ( $this->is_admin_creation_process() ) {
+				return;
+			}
+
 	}
 
 	/**
@@ -140,25 +143,29 @@ class UR_User_Approval {
 	 *
 	 * @param $user_id
 	 */
-	public function set_user_status( $user_id ) {
+	public function set_user_status( $form_data, $form_id, $user_id ) {
 
-		$status = UR_Admin_User_Manager::PENDING;
 
-		// If the user is created by admin in the backend, than automatically approve him
-		if ( $this->is_admin_creation_process() ) {
-			$status = UR_Admin_User_Manager::APPROVED;
+		if ( 'admin_approval' === ur_get_single_post_meta( $form_id, 'user_registration_form_setting_login_options', get_option( 'user_registration_general_setting_login_options', 'default' ) ) ) {
+
+			$status = UR_Admin_User_Manager::PENDING;
+
+			// If the user is created by admin in the backend, than automatically approve him
+			if ( $this->is_admin_creation_process() ) {
+				$status = UR_Admin_User_Manager::APPROVED;
+			}
+
+			$user_manager = new UR_Admin_User_Manager( $user_id );
+
+			// The user have to be not alerted on status creation, it will be always pending or approved
+			$alert_user = false;
+
+			$user_manager->save_status( $status, $alert_user );
 		}
-
-		$user_manager = new UR_Admin_User_Manager( $user_id );
-
-		// The user have to be not alerted on status creation, it will be always pending or approved
-		$alert_user = false;
-
-		$user_manager->save_status( $status, $alert_user );
 	}
 
 	/**
-	 * Check the status of an user on login
+	 * Check the status of an user on login.
 	 *
 	 * @param $user
 	 *
@@ -169,27 +176,64 @@ class UR_User_Approval {
 	 */
 	public function check_status_on_login( WP_User $user, $password ) {
 
-		$user_manager = new UR_Admin_User_Manager( $user );
 
-		$status = $user_manager->get_user_status();
+		$form_id       = ur_get_form_id_by_userid( $user->ID );
 
-		do_action( 'ur_user_before_check_status_on_login', $status, $user );
+		if ( 'admin_approval' === ur_get_single_post_meta( $form_id, 'user_registration_form_setting_login_options', get_option( 'user_registration_general_setting_login_options', 'default' ) ) ) {
+			$user_manager = new UR_Admin_User_Manager( $user );
 
-		switch ( $status ) {
-			case UR_Admin_User_Manager::APPROVED:
-				return $user;
-				break;
-			case UR_Admin_User_Manager::PENDING:
-				$message = '<strong>' . __( 'ERROR:', 'user-registration' ) . '</strong> ' . __( 'Your account is still pending approval.', 'user-registration' );
+			$status = $user_manager->get_user_status();
 
-				return new WP_Error( 'pending_approval', $message );
-				break;
-			case UR_Admin_User_Manager::DENIED:
-				$message = '<strong>' . __( 'ERROR:', 'user-registration' ) . '</strong> ' . __( 'Your account has been denied.', 'user-registration' );
+			do_action( 'ur_user_before_check_status_on_login', $status, $user );
 
-				return new WP_Error( 'denied_access', $message );
-				break;
+			switch ( $status ) {
+				case UR_Admin_User_Manager::APPROVED:
+					return $user;
+					break;
+				case UR_Admin_User_Manager::PENDING:
+					$message = '<strong>' . __( 'ERROR:', 'user-registration' ) . '</strong> ' . __( 'Your account is still pending approval.', 'user-registration' );
+
+					return new WP_Error( 'pending_approval', $message );
+					break;
+				case UR_Admin_User_Manager::DENIED:
+					$message = '<strong>' . __( 'ERROR:', 'user-registration' ) . '</strong> ' . __( 'Your account has been denied.', 'user-registration' );
+
+					return new WP_Error( 'denied_access', $message );
+					break;
+			}
+		} else if ( 'email_confirmation' === ur_get_single_post_meta( $form_id, 'user_registration_form_setting_login_options', get_option( 'user_registration_general_setting_login_options', 'default' ) ) ) {
+			$email_status = get_user_meta( $user->ID, 'ur_confirm_email', true );
+
+			do_action( 'ur_user_before_check_email_status_on_login', $email_status, $user );
+
+			$url = ( ! empty( $_SERVER['HTTPS'] ) ) ? 'https://' . $_SERVER['SERVER_NAME'] . $_SERVER['REQUEST_URI'] : 'http://' . $_SERVER['SERVER_NAME'] . $_SERVER['REQUEST_URI'];
+			$url = substr( $url, 0, strpos( $url, '?' ) );
+			$instance = new UR_Email_Confirmation();
+			$url = wp_nonce_url( $url . '?ur_resend_id=' . $instance->crypt_the_string( $user->ID, 'e' ) . '&ur_resend_token=true', 'ur_resend_token' );
+
+			if ( $email_status === '0' ) {
+				$message = '<strong>' . __( 'ERROR:', 'user-registration' ) . '</strong> ' . sprintf( __( 'Your account is still pending approval. Verify your email by clicking on the link sent to your email. %s', 'user-registration' ), '<a id="resend-email" href="' . esc_url( $url ) . '">' . __( 'Resend Verification Link', 'user-registration' ) . '</a>' );
+				return new WP_Error( 'user_email_not_verified', $message );
+			}
+			return $user;
+		} elseif ( 'payment' === ur_get_single_post_meta( $form_id, 'user_registration_form_setting_login_options', get_option( 'user_registration_general_setting_login_options', 'default' ) ) ) {
+			$payment_status = get_user_meta( $user->ID, 'ur_payment_status', true );
+
+			do_action( 'ur_user_before_check_payment_status_on_login', $payment_status, $user );
+
+			if( ! empty( $payment_status ) && $payment_status !== 'completed' ) {
+
+				$user_id = $user->ID;
+				$instance = new User_Registration_Payments_Process();
+				$redirect_url = $instance->generate_redirect_url( $user_id );
+				$message = '<strong>' . __( 'ERROR:', 'user-registration-payments' ) . '</strong> ' . sprintf( __( 'Your account is still pending payment. Process the payment by clicking on this: %s', 'user-registration-payments' ), '<a id="payment-link" href="'. esc_url( $redirect_url ) .'">'. __( 'link', 'user-registration-payments' ). '</a>' );
+
+				return new WP_Error( 'user_payment_pending', $message );
+			}
+
+			return $user;
 		}
+		return $user;
 	}
 
 	/**
@@ -202,21 +246,25 @@ class UR_User_Approval {
 			return;
 		}
 
-		$status = ur_get_user_approval_status( get_current_user_id() );
+		$form_id       = ur_get_form_id_by_userid( get_current_user_id() );
 
-		$user_manager = new UR_Admin_User_Manager();
+		if ( 'admin_approval' === ur_get_single_post_meta( $form_id, 'user_registration_form_setting_login_options', get_option( 'user_registration_general_setting_login_options', 'default' ) ) ) {
+			$status = ur_get_user_approval_status( get_current_user_id() );
 
-		if ( ! $user_manager->can_status_be_changed_by( get_current_user_id() ) ) {
-			return;
+			$user_manager = new UR_Admin_User_Manager();
+
+			if ( ! $user_manager->can_status_be_changed_by( get_current_user_id() ) ) {
+				return;
+			}
+
+			do_action( 'ur_user_before_check_status_on_page', $status, $user_manager );
+
+			if ( $status == UR_Admin_User_Manager::APPROVED ) {
+				return;
+			}
+
+			wp_logout();
 		}
-
-		do_action( 'ur_user_before_check_status_on_page', $status, $user_manager );
-
-		if ( $status == UR_Admin_User_Manager::APPROVED ) {
-			return;
-		}
-
-		wp_logout();
 	}
 
 	/**
@@ -234,12 +282,15 @@ class UR_User_Approval {
 	 * @param $user_id
 	 */
 	public function disconnect_user_session( $user_id ) {
+		$form_id       = ur_get_form_id_by_userid( $user_id );
 
-		// get all sessions for user with ID $user_id
-		$sessions = WP_Session_Tokens::get_instance( $user_id );
+		if ( 'admin_approval' === ur_get_single_post_meta( $form_id, 'user_registration_form_setting_login_options', get_option( 'user_registration_general_setting_login_options', 'default' ) ) ) {
+			// get all sessions for user with ID $user_id
+			$sessions = WP_Session_Tokens::get_instance( $user_id );
 
-		// we have got the sessions, destroy them all!
-		$sessions->destroy_all();
+			// we have got the sessions, destroy them all!
+			$sessions->destroy_all();
+		}
 	}
 
 	/**
@@ -252,11 +303,15 @@ class UR_User_Approval {
 	 */
 	public function allow_password_reset( $result, $user_id ) {
 
-		$user_manager = new UR_Admin_User_Manager( $user_id );
+		$form_id       = ur_get_form_id_by_userid( $user_id );
 
-		if ( ! $user_manager->is_approved() ) {
-			$error_message = __( 'Your account is still awaiting admin approval. Reset Password is not allowed.', 'user-registration' );
-			$result        = new WP_Error( 'user_not_approved', $error_message );
+		if ( 'admin_approval' === ur_get_single_post_meta( $form_id, 'user_registration_form_setting_login_options', get_option( 'user_registration_general_setting_login_options', 'default' ) ) ) {
+			$user_manager = new UR_Admin_User_Manager( $user_id );
+
+			if ( ! $user_manager->is_approved() ) {
+				$error_message = __( 'Your account is still awaiting admin approval. Reset Password is not allowed.', 'user-registration' );
+				$result        = new WP_Error( 'user_not_approved', $error_message );
+			}
 		}
 
 		return $result;
